- Added test cases.
