+++ clearblade-cloud-iot/pom.xml
@@ -214,5 +214,11 @@
 			<artifactId>guava</artifactId>
 			<version>31.1-jre</version>
 		</dependency>
+		<dependency>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter-api</artifactId>
+			<version>5.8.2</version>
+			<scope>compile</scope>
+		</dependency>
 	</dependencies>
 </project>
